pyproject.toml
[project]
name = "schnapsen-llm-bench"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksey Martynyuk", email = "a.m.martynyuk@student.vu.nl"},
    {name = "Tommaso Zambelli Franz", email = "t.zambellifranz@student.vu.nl"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = "^3.10"
schnapsen = "0.0.5"
openai = "^1.58.0"
poetry = "^2.0.0"
pydantic = "^2.8.0"
psycopg = {extras = ["binary"], version = "^3.2.4"}
azure-cosmos = "^4.9.0"
python-dotenv = "^1.0.0"
jupyter = "^1.1.1"
pandas = "^2.2.3"
azure-identity = "^1.19.0"
mistralai = "^1.4.0"
seaborn = "^0.13.0"
matplotlib = "^3.8.0"
tabulate = "^0.9.0"
statsmodels = "^0.14.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"